requirements.txt
# Core
google-adk>=0.4.0
google-generativeai>=0.8.0

# Agent Tools & APIs
yfinance>=0.2.40
tavily-python>=0.3.0
newsapi-python>=0.2.7
wikipedia>=1.4.0
requests>=2.32.0
feedparser>=6.0.11
praw>=7.8.0

# MCP
mcp>=1.0.0

# Frontend
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
asyncio-throttle>=1.0.2
tenacity>=9.0.0
rich>=13.9.0

# A2A
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.32.0
